added cumulative transforms
